Renamed React files to .jsx
